+++ cards/helpers/all.js
@@ -1,0 +1,77 @@
+const BasicShieldCard = require('../basic-shield');
+const BattleFocusCard = require('../battle-focus');
+const BerserkCard = require('../berserk');
+const BlastCard = require('../blast');
+const BlinkCard = require('../blink');
+const BoostCard = require('../boost');
+const BrainDrainCard = require('../brain-drain');
+const CamouflageVestCard = require('../camouflage-vest');
+const CloakOfInvisibilityCard = require('../cloak-of-invisibility');
+const CoilCard = require('../coil');
+const ConstrictCard = require('../constrict');
+const CurseCard = require('../curse');
+const EnchantedFaceswapCard = require('../enchanted-faceswap');
+const EnthrallCard = require('../enthrall');
+const EntranceCard = require('../entrance');
+const FightOrFlightCard = require('../fight-or-flight');
+const FistsOfVillainyCard = require('../fists-of-villainy');
+const FistsOfVirtueCard = require('../fists-of-virtue');
+const ForkedMetalRodCard = require('../forked-metal-rod');
+const ForkedStickCard = require('../forked-stick');
+const FleeCard = require('../flee');
+const HealCard = require('../heal');
+const HitCard = require('../hit');
+const HitHarderCard = require('../hit-harder');
+const HornGoreCard = require('../horn-gore');
+const KalevalaCard = require('../kalevala');
+const PickPocketCard = require('../pick-pocket');
+const LuckyStrike = require('../lucky-strike');
+const MesmerizeCard = require('../mesmerize');
+const PoundCard = require('../pound');
+const RandomCard = require('../random');
+const RehitCard = require('../rehit');
+const ScotchCard = require('../scotch');
+const ThickSkinCard = require('../thick-skin');
+const VenegefulRampageCard = require('../vengeful-rampage');
+const WhiskeyShotCard = require('../whiskey-shot');
+const WoodenSpearCard = require('../wooden-spear');
+
+module.exports = [
+	BasicShieldCard,
+	BattleFocusCard,
+	BerserkCard,
+	BlastCard,
+	BlinkCard,
+	BoostCard,
+	BrainDrainCard,
+	CamouflageVestCard,
+	CloakOfInvisibilityCard,
+	CoilCard,
+	ConstrictCard,
+	CurseCard,
+	EnchantedFaceswapCard,
+	EnthrallCard,
+	EntranceCard,
+	FightOrFlightCard,
+	FistsOfVillainyCard,
+	FistsOfVirtueCard,
+	FleeCard,
+	ForkedMetalRodCard,
+	ForkedStickCard,
+	HealCard,
+	HitCard,
+	HitHarderCard,
+	HornGoreCard,
+	KalevalaCard,
+	PickPocketCard,
+	LuckyStrike,
+	MesmerizeCard,
+	PoundCard,
+	RandomCard,
+	RehitCard,
+	ScotchCard,
+	ThickSkinCard,
+	VenegefulRampageCard,
+	WhiskeyShotCard,
+	WoodenSpearCard
+];
