+++ cards/helpers/constants.js
@@ -1,0 +1,3 @@
+module.exports = {
+	DEFAULT_MINIMUM_CARDS: 14
+};
