+++ cards/helpers/sort.js
@@ -1,0 +1,11 @@
+module.exports = cards => cards.sort((a, b) => {
+	if (a.cardType > b.cardType) {
+		return 1;
+	}
+
+	if (a.cardType < b.cardType) {
+		return -1;
+	}
+
+	return 0;
+});
