+++ cards/helpers/unique-cards.js
@@ -1,0 +1,6 @@
+module.exports = cards => cards.reduce(
+	(uniqueCards, card) =>
+		uniqueCards.concat(!uniqueCards.find(possibleCard =>
+			possibleCard.name === card.name) ? [card] : [])
+	, []
+);
